@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { getSmurfs, addSmurf } from '../actions';
 import './App.css';
@@ -8,35 +8,33 @@ import './App.css';
  Just remember, `how do I `connect` my components to redux?`
  `How do I ensure that my component links the state to props?`
  */
-class App extends Component {
-  componentDidMount() {
-    this.props.getSmurfs();
-  }
-  addSmurf = (event) => {
+const App = ({smurfs, ...props}) => {
+  useEffect(() => props.getSmurfs(), []);
+  const performAddSmurf = (event) => {
     event.persist();
     event.preventDefault();
     const name = event.target.name.value.trim();
     const age = parseInt(event.target.age.value, 10);
     const height = event.target.height.value.trim();
     if (name !== "" && age && height !=="") {
-      this.props.addSmurf({name, age, height});
+      props.addSmurf({name, age, height});
     }
   };
-  render() {
-    return (
-      <div className="App">
-        <h1>SMURFS! 2.0 W/ Redux</h1>
-        {this.props.smurfs.map(smurf => <div key={smurf.id}>{smurf.name}, {smurf.age}, {smurf.height}</div>)}
-        <form onSubmit={this.addSmurf}>
-          <input type="text" name="name" placeholder="name" />
-          <input type="number" name="age" placeholder="age" />
-          <input type="text" name="height" placeholder="height" />
-          <button type="submit">Add Smurf</button>
-        </form>
-      </div>
-    );
-  }
-}
+  return (
+    <div className="App">
+      <h1>SMURFS! 2.0 W/ Redux</h1>
+      {smurfs.map(smurf => (
+        <div key={smurf.id}>{smurf.name}, {smurf.age}, {smurf.height}</div>
+      ))}
+      <form onSubmit={performAddSmurf}>
+        <input type="text" name="name" placeholder="name" />
+        <input type="number" name="age" placeholder="age" />
+        <input type="text" name="height" placeholder="height" />
+        <button type="submit">Add Smurf</button>
+      </form>
+    </div>
+  );
+};
 
 const mapStateToProps = ({gettingSmurfs, smurfs}) => ({ gettingSmurfs, smurfs });
 
